requirements.txt
# Facebook Marketplace Agent - Professional Edition Requirements
# Core dependencies for professional automation

# Web automation and browser control
playwright>=1.40.0
beautifulsoup4>=4.12.2

# Image processing and manipulation
Pillow>=10.1.0

# HTTP requests and web scraping
requests>=2.31.0
aiohttp>=3.9.0

# Data processing and utilities
pandas>=2.1.0
numpy>=1.24.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging and monitoring
loguru>=0.7.2

# File and path utilities
pathlib2>=2.3.7

# Text processing and natural language
nltk>=3.8.1
fuzzy>=1.2.2

# Date and time utilities
python-dateutil>=2.8.2

# Async utilities
asyncio-throttle>=1.0.2

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Professional features
# Advanced form handling
selenium>=4.15.0  # Fallback browser automation

# Image recognition and processing
opencv-python>=4.8.0

# Machine learning for behavior simulation
scikit-learn>=1.3.0

# Database support (optional)
sqlalchemy>=2.0.0
alembic>=1.12.0

# API and web framework support
fastapi>=0.104.0
uvicorn>=0.24.0

# Monitoring and metrics
prometheus-client>=0.19.0

# Security and encryption
cryptography>=41.0.0

# Professional deployment
gunicorn>=21.2.0
nginx-python>=1.1.0

# Note: Install Playwright browsers after pip install
# Run: playwright install chromium